Get clients built on darwin.

+++ tests/mtstressnast.c
@@ -43,7 +43,7 @@ io_looper(void *arg)
 				after.tv_usec += 1000000;
 			}
 			fprintf(stderr, "Request time: %2ld.%06ld seconds.\n",
-				after.tv_sec, after.tv_usec);
+				after.tv_sec, (unsigned long)after.tv_usec);
 		}
 	}
 }
